Update ControllerManager
Renderer section of the script changes the sphere color from red to blue when the controller is in different areas of the scene. This could be used when there are two bodies in one scene, where the body on the left is colored red and the body on the right is colored blue; the spheres would automatically be red for the body on the left and blue for the body on the right. May have to work around the 3D aspects, where people's orientation will change and they could get confused about which body (activated or deactivated) they are coloring.