Create FileCreator
Need a folder in assets called "TextFiles". Creates files for the folder each run with the names listed in nameList. Allows you to do run a trial of one patient without having to touch the files. After going through all the emotions with one patient, save the text files on your computer, and remove them from the TextFiles folder.